changes in wp, not generated yet

+++ wp-html/wp-content/plugins/wpforms-lite/includes/providers/class-constant-contact.php
@@ -11,6 +11,7 @@ class WPForms_Constant_Contact extends WPForms_Provider {
 	 * Provider access token.
 	 *
 	 * @since 1.3.6
+     *
 	 * @var string
 	 */
 	public $access_token;
@@ -19,6 +20,7 @@ class WPForms_Constant_Contact extends WPForms_Provider {
 	 * Provider API key.
 	 *
 	 * @since 1.3.6
+     *
 	 * @var string
 	 */
 	public $api_key = 'c58xq3r27udz59h9rrq7qnvf';
@@ -27,6 +29,7 @@ class WPForms_Constant_Contact extends WPForms_Provider {
 	 * Sign up link.
 	 *
 	 * @since 1.3.6
+     *
 	 * @var string
 	 */
 	public $sign_up = 'https://constant-contact.evyy.net/c/11535/341874/3411?sharedid=wpforms';
@@ -46,12 +49,15 @@ class WPForms_Constant_Contact extends WPForms_Provider {
 
 		if ( is_admin() ) {
 			// Admin notice requesting connecting.
-			add_action( 'admin_notices', array( $this, 'connect_request' ) );
-			add_action( 'wp_ajax_wpforms_constant_contact_dismiss', array( $this, 'connect_dismiss' ) );
-			add_action( 'wpforms_admin_page', array( $this, 'learn_more_page' ) );
+			$this->connect_request();
+
+			add_action( 'wpforms_admin_notice_dismiss_ajax', [ $this, 'connect_dismiss' ] );
+			add_action( 'wpforms_admin_page', [ $this, 'learn_more_page' ] );
+			add_filter( "wpforms_providers_provider_settings_formbuilder_display_content_default_screen_{$this->slug}", [ $this, 'builder_settings_default_content' ] );
 
 			// Provide option to override sign up link.
 			$sign_up = get_option( 'wpforms_constant_contact_signup', false );
+
 			if ( $sign_up ) {
 				$this->sign_up = esc_html( $sign_up );
 			}
@@ -66,7 +72,7 @@ class WPForms_Constant_Contact extends WPForms_Provider {
 	 * @param array $fields    List of fields with their data and settings.
 	 * @param array $entry     Submitted entry values.
 	 * @param array $form_data Form data and settings.
-	 * @param int $entry_id    Saved entry ID.
+	 * @param int   $entry_id  Saved entry ID.
 	 *
 	 * @return void
 	 */
@@ -516,64 +522,48 @@ class WPForms_Constant_Contact extends WPForms_Provider {
 	 *
 	 * @since 1.3.6
 	 *
-	 * @param string $connection_id
-	 * @param array $connection
+	 * @param string $connection_id Connection Id.
+	 * @param array  $connection    Connection data.
 	 *
 	 * @return string
 	 */
-	public function output_groups( $connection_id = '', $connection = array() ) {
+	public function output_groups( $connection_id = '', $connection = [] ) {
+
 		// No groups or segments for this provider.
 		return '';
 	}
 
 	/**
-	 * Output content after the main builder output.
+	 * Default content for the provider settings panel in the form builder.
 	 *
-	 * @since 1.3.6
+	 * @since 1.6.8
+	 *
+	 * @param string $content Default content.
+	 *
+	 * @return string
 	 */
-	public function builder_output_after() {
+	public function builder_settings_default_content( $content ) {
 
-		// Only display if Constant Contact account has not been setup.
-		$providers = wpforms_get_providers_options();
-
-		if ( ! empty( $providers[ $this->slug ] ) ) {
-			return;
-		}
+		ob_start();
 		?>
-		<div class="wpforms-alert wpforms-alert-info">
-			<p>
-				<?php
-				echo wp_kses(
-					__( 'Get the most out of <strong>WPForms</strong> &mdash; use it with an active Constant Contact account.', 'wpforms-lite' ),
-					array(
-						'strong' => array(),
-					)
-				);
-				?>
-			</p>
-			<p>
-				<a href="<?php echo esc_url( $this->sign_up ); ?>" style="margin-right: 10px;" class="button-primary" target="_blank" rel="noopener noreferrer">
-					<?php esc_html_e( 'Try Constant Contact for Free', 'wpforms-lite' ); ?>
-				</a>
-				<?php
-				printf(
-					wp_kses(
-						/* translators: %s - WPForms Constant Contact internal URL. */
-						__( 'Learn More about the <a href="%s" target="_blank" rel="noopener noreferrer">power of email marketing</a>', 'wpforms-lite' ),
-						array(
-							'a' => array(
-								'href'   => array(),
-								'target' => array(),
-								'rel'    => array(),
-							),
-						)
-					),
-					esc_url( admin_url( 'admin.php?page=wpforms-page&wpforms-page=constant-contact' ) )
-				);
-				?>
-			</p>
-		</div>
+		<p>
+			<a href="<?php echo esc_url( $this->sign_up ); ?>" class="wpforms-btn wpforms-btn-md wpforms-btn-orange" target="_blank" rel="noopener noreferrer">
+				<?php esc_html_e( 'Try Constant Contact for Free', 'wpforms-lite' ); ?>
+			</a>
+		</p>
+		<p>
+			<?php
+			printf(
+				'<a href="%s" target="_blank" rel="noopener noreferrer" class="secondary-text">' .
+					esc_html__( 'Learn more about the power of email marketing.', 'wpforms-lite' ) .
+				'</a>',
+				esc_url( admin_url( 'admin.php?page=wpforms-page&wpforms-page=constant-contact' ) )
+			);
+			?>
+		</p>
 		<?php
+
+		return $content . ob_get_clean();
 	}
 
 	/*************************************************************************
@@ -625,17 +615,18 @@ class WPForms_Constant_Contact extends WPForms_Provider {
 		}
 
 		// Don't display on WPForms admin content pages.
-		if ( ! empty( $_GET['wpforms-page'] ) ) {
+		if ( ! empty( $_GET['wpforms-page'] ) ) { // phpcs:ignore WordPress.Security.NonceVerification.Recommended
 			return;
 		}
 
 		// Don't display if user is about to connect via Settings page.
-		if ( ! empty( $_GET['wpforms-integration'] ) && $this->slug === $_GET['wpforms-integration'] ) {
+		if ( ! empty( $_GET['wpforms-integration'] ) && $this->slug === $_GET['wpforms-integration'] ) { // phpcs:ignore WordPress.Security.NonceVerification.Recommended
 			return;
 		}
 
-		// Only display the notice is the Constant Contact option is set and
+		// Only display the notice if the Constant Contact option is set and
 		// there are previous Constant Contact connections created.
+		// Please do not delete 'wpforms_constant_contact' option check from the code.
 		$cc_notice = get_option( 'wpforms_constant_contact', false );
 		$providers = wpforms_get_providers_options();
 
@@ -646,42 +637,42 @@ class WPForms_Constant_Contact extends WPForms_Provider {
 		// Output the notice message.
 		$connect    = admin_url( 'admin.php?page=wpforms-settings&wpforms-integration=constant-contact#!wpforms-tab-providers' );
 		$learn_more = admin_url( 'admin.php?page=wpforms-page&wpforms-page=constant-contact' );
+
+		ob_start();
 		?>
-		<div class="notice notice-info is-dismissible wpforms-constant-contact-notice">
-			<p>
-				<?php
-				echo wp_kses(
-					__( 'Get the most out of the <strong>WPForms</strong> plugin &mdash; use it with an active Constant Contact account.', 'wpforms-lite' ),
-					array(
-						'strong' => array(),
-					)
-				);
-				?>
-			</p>
-			<p>
-				<a href="<?php echo esc_url( $this->sign_up ); ?>" class="button-primary" target="_blank" rel="noopener noreferrer">
-					<?php esc_html_e( 'Try Constant Contact for Free', 'wpforms-lite' ); ?>
-				</a>
-				<a href="<?php echo esc_url( $connect ); ?>" class="button-secondary">
-					<?php esc_html_e( 'Connect your existing account', 'wpforms-lite' ); ?>
-				</a>
-				<?php
-				printf(
-					wp_kses(
-						/* translators: %s - WPForms Constant Contact internal URL. */
-						__( 'Learn More about the <a href="%s">power of email marketing</a>', 'wpforms-lite' ),
-						array(
-							'a' => array(
-								'href' => array(),
-							),
-						)
-					),
-					esc_url( $learn_more )
-				);
-				?>
-			</p>
-		</div>
-		<style type="text/css">
+		<p>
+			<?php
+			echo wp_kses(
+				__( 'Get the most out of the <strong>WPForms</strong> plugin &mdash; use it with an active Constant Contact account.', 'wpforms-lite' ),
+				[
+					'strong' => [],
+				]
+			);
+			?>
+		</p>
+		<p>
+			<a href="<?php echo esc_url( $this->sign_up ); ?>" class="button-primary" target="_blank" rel="noopener noreferrer">
+				<?php esc_html_e( 'Try Constant Contact for Free', 'wpforms-lite' ); ?>
+			</a>
+			<a href="<?php echo esc_url( $connect ); ?>" class="button-secondary">
+				<?php esc_html_e( 'Connect your existing account', 'wpforms-lite' ); ?>
+			</a>
+			<?php
+			printf(
+				wp_kses( /* translators: %s - WPForms Constant Contact internal URL. */
+					__( 'Learn More about the <a href="%s">power of email marketing</a>', 'wpforms-lite' ),
+					[
+						'a' => [
+							'href' => [],
+						],
+					]
+				),
+				esc_url( $learn_more )
+			);
+			?>
+		</p>
+
+		<style>
 			.wpforms-constant-contact-notice {
 				border-left-color: #1a5285;
 			}
@@ -700,26 +691,37 @@ class WPForms_Constant_Contact extends WPForms_Provider {
 				margin: 0 10px 0 0;
 			}
 		</style>
-		<script type="text/javascript">
-			jQuery( function ( $ ) {
-				$( document ).on( 'click', '.wpforms-constant-contact-notice button', function ( event ) {
-					event.preventDefault();
-					$.post( ajaxurl, { action: 'wpforms_constant_contact_dismiss' } );
-					$( '.wpforms-constant-contact-notice' ).remove();
-				} );
-			} );
-		</script>
 		<?php
+
+		$notice = ob_get_clean();
+
+		\WPForms\Admin\Notice::info(
+			$notice,
+			[
+				'dismiss' => \WPForms\Admin\Notice::DISMISS_GLOBAL,
+				'slug'    => 'constant_contact_connect',
+				'autop'   => false,
+				'class'   => 'wpforms-constant-contact-notice',
+			]
+		);
 	}
 
 	/**
 	 * Dismiss the Constant Contact admin notice.
 	 *
 	 * @since 1.3.6
+	 * @since 1.6.7.1 Added parameter $notice_id.
+	 *
+	 * @param string $notice_id Notice ID (slug).
 	 */
-	public function connect_dismiss() {
+	public function connect_dismiss( $notice_id = '' ) {
+
+		if ( $notice_id !== 'global-constant_contact_connect' ) {
+			return;
+		}
 
 		delete_option( 'wpforms_constant_contact' );
+
 		wp_send_json_success();
 	}
 
@@ -824,7 +826,8 @@ class WPForms_Constant_Contact extends WPForms_Provider {
 			<p><?php esc_html_e( 'It doesn&#39;t matter what kind of business you run, what kind of website you have, or what industry you are in - you need to start building your email list today.', 'wpforms-lite' ); ?></p>
 			<p><?php esc_html_e( 'With Constant Contact + WPForms, growing your list is easy.', 'wpforms-lite' ); ?></p>
 		</div>
-		<style type="text/css">
+
+		<style>
 			.notice {
 				display: none;
 			}
@@ -971,4 +974,4 @@ class WPForms_Constant_Contact extends WPForms_Provider {
 	}
 }
 
-new WPForms_Constant_Contact;
+new WPForms_Constant_Contact();
